fix: events with `groupId` set to `0` won't show
